Remove all message interpretation from worker to server
Per vdbergh's comments

Still untested
